+++ src/commander/log.rs
@@ -6,6 +6,7 @@ It is mostly used in the [log_tab][crate::ui::log_tab] module.
 */
 
 use std::fmt::Display;
+use std::process::Child;
 use std::sync::LazyLock;
 
 use anyhow::Context;
@@ -19,6 +20,7 @@ use tracing::instrument;
 
 use crate::commander::CommandError;
 use crate::commander::Commander;
+use crate::commander::JjCommand;
 use crate::commander::RemoveEndLine;
 use crate::commander::bookmarks::Bookmark;
 use crate::commander::ids::ChangeId;
@@ -160,22 +162,35 @@ impl Commander {
         })
     }
 
-    /// Get commit details.
+    /// Spawn child process to get commit details.
     /// Maps to `jj show <commit>`
     #[instrument(level = "trace", skip(self))]
-    pub fn get_commit_show(
+    pub fn spawn_commit_show(
         &self,
         commit_id: &CommitId,
         diff_format: &DiffFormat,
         ignore_working_copy: bool,
-    ) -> Result<String, CommandError> {
+    ) -> Result<Child, CommandError> {
+        self.build_jj_commit_show(commit_id, diff_format, ignore_working_copy)
+            .color()
+            .spawn()
+    }
+
+    /// Create the JjCommmand for `jj show <commit>`
+    #[instrument(level = "trace", skip(self))]
+    pub fn build_jj_commit_show(
+        &self,
+        commit_id: &CommitId,
+        diff_format: &DiffFormat,
+        ignore_working_copy: bool,
+    ) -> JjCommand<'_> {
         let mut args = vec!["show", commit_id.as_str()];
         args.append(&mut diff_format.get_args());
         if ignore_working_copy {
             args.push("--ignore-working-copy");
         }
 
-        Ok(self.jj(args).color().run()?.remove_end_line())
+        self.jj(args)
     }
 
     /// Get the current head.
@@ -325,16 +340,17 @@ mod tests {
     }
 
     #[test]
-    fn get_commit_show() -> Result<()> {
+    fn spawn_commit_show() -> Result<()> {
         let test_repo = TestRepo::new()?;
 
         fs::write(test_repo.directory.path().join("README"), b"AAA")?;
 
         let head = test_repo.commander.get_current_head()?;
-        let show =
-            test_repo
-                .commander
-                .get_commit_show(&head.commit_id, &DiffFormat::ColorWords, false)?;
+        let output = test_repo
+            .commander
+            .spawn_commit_show(&head.commit_id, &DiffFormat::ColorWords, false)?
+            .wait_with_output()?;
+        let show = String::from_utf8(output.stdout)?.remove_end_line();
 
         let mut settings = insta::Settings::clone_current();
         settings.add_filter(r"Commit ID: [0-9a-fA-F]{40}", "Commit ID: [COMMIT_ID]");